pyproject.toml
[tool.poetry]
name = "store_core"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = { version = "^0.18.3", extras = ["standard"] }
pydantic = {version = "^1.10.2", extras = ["dotenv"]}
yarl = "^1.8.1"
ujson = "^5.5.0"
SQLAlchemy = {version = "^1.4.41", extras = ["mypy", "asyncio"]}
alembic = "^1.8.1"
asyncpg = {version = "^0.26.0", extras = ["sa"]}
httptools = "^0.5.0"
strawberry-graphql = { version = "^0.133.2", extras = ["fastapi"] }
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-exporter-otlp = "^1.15.0"
opentelemetry-instrumentation = "^0.36b0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-logging = "^0.36b0"
opentelemetry-instrumentation-sqlalchemy = "^0.36b0"
typeguard = "^4.1.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
flake8 = "~4.0.1"
mypy = "^0.981"
isort = "^5.10.1"
yesqa = "^1.4.0"
pre-commit = "^2.20.0"
wemake-python-styleguide = "^0.17.0"
black = "^22.8.0"
autoflake = "^1.6.1"
SQLAlchemy = {version = "^1.4.41", extras = ["mypy"]}
pytest-cov = "^4.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
httpx = "^0.23.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["store_core",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "STORE_CORE_DB_BASE=store_core_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
